[fix][store] Refactor txn api integer type.

+++ src/engine/engine.h
@@ -136,15 +136,15 @@ class Engine {
     TxnReader() = default;
     virtual ~TxnReader() = default;
 
-    virtual butil::Status TxnBatchGet(std::shared_ptr<Context> ctx, uint64_t start_ts,
+    virtual butil::Status TxnBatchGet(std::shared_ptr<Context> ctx, int64_t start_ts,
                                       const std::vector<std::string>& keys, std::vector<pb::common::KeyValue>& kvs,
                                       pb::store::TxnResultInfo& txn_result_info) = 0;
-    virtual butil::Status TxnScan(std::shared_ptr<Context> ctx, uint64_t start_ts, const pb::common::Range& range,
-                                  uint64_t limit, bool key_only, bool is_reverse,
+    virtual butil::Status TxnScan(std::shared_ptr<Context> ctx, int64_t start_ts, const pb::common::Range& range,
+                                  int64_t limit, bool key_only, bool is_reverse,
                                   pb::store::TxnResultInfo& txn_result_info, std::vector<pb::common::KeyValue>& kvs,
                                   bool& has_more, std::string& end_key) = 0;
-    virtual butil::Status TxnScanLock(std::shared_ptr<Context> ctx, uint64_t min_lock_ts, uint64_t max_lock_ts,
-                                      const pb::common::Range& range, uint64_t limit,
+    virtual butil::Status TxnScanLock(std::shared_ptr<Context> ctx, int64_t min_lock_ts, int64_t max_lock_ts,
+                                      const pb::common::Range& range, int64_t limit,
                                       std::vector<pb::store::LockInfo>& lock_infos) = 0;
   };
 
@@ -162,9 +162,9 @@ class Engine {
     virtual butil::Status TxnPrewrite(std::shared_ptr<Context> ctx, const std::vector<pb::store::Mutation>& mutations,
                                       const std::string& primary_lock, int64_t start_ts, int64_t lock_ttl,
                                       int64_t txn_size, bool try_one_pc, int64_t max_commit_ts,
-                                      const std::vector<int32_t>& pessimistic_checks,
-                                      const std::map<int32_t, int64_t>& for_update_ts_checks,
-                                      const std::map<int32_t, std::string>& lock_extra_datas) = 0;
+                                      const std::vector<int64_t>& pessimistic_checks,
+                                      const std::map<int64_t, int64_t>& for_update_ts_checks,
+                                      const std::map<int64_t, std::string>& lock_extra_datas) = 0;
     virtual butil::Status TxnCommit(std::shared_ptr<Context> ctx, int64_t start_ts, int64_t commit_ts,
                                     const std::vector<std::string>& keys) = 0;
     virtual butil::Status TxnCheckTxnStatus(std::shared_ptr<Context> ctx, const std::string& primary_key,
