+++ src/engine/txn_engine_helper.h
@@ -35,7 +35,7 @@ class TxnEngineHelper {
   static butil::Status GetLockInfo(RawEngine::ReaderPtr reader, const std::string &key, pb::store::LockInfo &lock_info);
 
   static butil::Status ScanLockInfo(RawEnginePtr raw_engine, int64_t min_lock_ts, int64_t max_lock_ts,
-                                    const std::string &start_key, const std::string &end_key, uint32_t limit,
+                                    const std::string &start_key, const std::string &end_key, int64_t limit,
                                     std::vector<pb::store::LockInfo> &lock_infos);
 
   static butil::Status BatchGet(RawEnginePtr raw_engine, const pb::store::IsolationLevel &isolation_level,
@@ -84,9 +84,9 @@ class TxnEngineHelper {
   static butil::Status Prewrite(RawEnginePtr raw_engine, std::shared_ptr<Engine> raft_engine,
                                 std::shared_ptr<Context> ctx, const std::vector<pb::store::Mutation> &mutations,
                                 const std::string &primary_lock, int64_t start_ts, int64_t lock_ttl, int64_t txn_size,
-                                bool try_one_pc, int64_t max_commit_ts, const std::vector<int32_t> &pessimistic_checks,
-                                const std::map<int32_t, int64_t> &for_update_ts_checks,
-                                const std::map<int32_t, std::string> &lock_extra_datas);
+                                bool try_one_pc, int64_t max_commit_ts, const std::vector<int64_t> &pessimistic_checks,
+                                const std::map<int64_t, int64_t> &for_update_ts_checks,
+                                const std::map<int64_t, std::string> &lock_extra_datas);
 
   static butil::Status Commit(RawEnginePtr raw_engine, std::shared_ptr<Engine> engine, std::shared_ptr<Context> ctx,
                               int64_t start_ts, int64_t commit_ts, const std::vector<std::string> &keys);
